document-scanner/requirements.txt
pytesseract
Pillow
transformers
watchdog
pdf2image
torch
urllib3
pyOpenSSL